autotest fixes

+++ internal/app/models/withdrawal.go
@@ -7,7 +7,7 @@ import (
 
 type Withdrawal struct {
 	Order       string    `json:"order"`
-	Sum         int       `json:"sum"`
+	Sum         float64   `json:"sum"`
 	ProcessedAt time.Time `json:"processed_at"`
 }
